requirements.txt
ariadne
broadcaster
websockets
asyncio_redis
uvicorn
starlette
python-multipart
ariadne[file-uploads]

requests
beautifulsoup4
aiodataloader
httpx
ariadne-graphql-modules
python-dotenv
python-json-logger
starlette-testclient
pytest
redis[hiredis]
pyjwt

opentelemetry-exporter-jaeger
opentelemetry-distro

opentelemetry-api 
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-django
opentelemetry-instrumentation-celery

gunicorn==20.1.0
psutil
asgi-background